Data layer and factory design pattern implementation

+++ src/main/kotlin/com/pawelcz/investments/calculationAlgorithm/Algorithm.kt
@@ -1,12 +1,20 @@
 package com.pawelcz.investments.calculationAlgorithm
 
+
 import com.pawelcz.investments.investment.Investment
+import java.math.BigDecimal
+
+
 
 abstract class Algorithm(
     var name : String
+
 ) {
 
-    abstract fun calculation(investment: Investment) : Double
+
+    abstract fun calculation(investment: Investment) : BigDecimal
+
+
